Gemstones Question
There is a collection of rocks where each rock has various minerals embeded in
it. Each type of mineral is designated by a lowercase letter in the range
ascii|a 2]. There may be multiple occurrences of a mineral in a rock. A
mineral is called a gemstone if it occurs at least once in each of the rocks in the
collection.
Given a list of minerals embedded in each of the rocks, display the number of
types of gemstones in the collection.

Example
arr = l'abe', 'abc', 'bc']
The minerals b and c appear in each rock, SO there are 2 gemstones.

Function Description
Complete the gemstones function in the editor below.
gemstones has the following parameter(s):
string arr[n]: an array of strings

Returns
int: the number of gemstones found

Input Format
The first line consists of an integer n, the size of arr.
Each of the next n lines contains a string arr li] where each letter represents an
occurence of a mineral in the current rock.