Alternating Characters Question
You are given a string containing characters A and B only. Your task is to
change it into a string such that there are no matching adjacent characters. To
do this, you are allowed to delete zero or more characters in the string.
Your task is to find the minimum number of required deletions.

Example
§ = A ABAAB
Remove an A at positions 0 and 3 to make S = A.BAB in 2 deletions.

Function Description
Complete the alternatingCharacters function in the editor below.
alternatingCharacters has the following parameter(s):
string s: a string

Returns
int: the minimum number of deletions required

Input Format
The first line contains an integer q, the number of queries.
The next g lines each contain a string s to analyze.